new model with 7 features
Added a new python file (7features.py) that is the same as 4features.py but also includes the 3 additional features 'WBC', 'ANC' and 'MONOCYTES'.

Also added a small section to check if the largest time value in the train set is larger than the largest one in the test set.